+++ app/src/main/java/edu/ucsb/cs/cs185/contenant/contenant/Room.java
@@ -10,17 +10,17 @@ public class Room implements Serializable {
     private static AtomicInteger nextId = new AtomicInteger();
 
     private final int id;
-    private final int houseId;
+    private final long houseId;
     private String notes;
     private int typeIndex; // The index in the room_array
 
-    public Room(int houseId, String notes) {
+    public Room(long houseId, String notes) {
         this.id = nextId.incrementAndGet();
         this.houseId = houseId;
         this.notes = notes;
     }
 
-    public Room(int houseId) {
+    public Room(long houseId) {
         this.id = nextId.incrementAndGet();
         this.houseId = houseId;
     }
@@ -37,7 +37,7 @@ public class Room implements Serializable {
         this.notes = notes;
     }
 
-    public int getHouseId() {
+    public long getHouseId() {
         return houseId;
     }
